Refactoring and add __init__.py
