+++ src/__init__.py
@@ -1,0 +1,2 @@
+__author__ = "Noah Ruben noah.ruben@online.de"
+__version__ = "Alpha v0.0.0.1"
